Restructure project: Converted loose collection of Python scripts into an installable package
