+++ .gitignore
@@ -8,3 +8,6 @@ venv/
 # Ignore Python cache files 
 __pycache__/ 
 *.pyc 
+
+# Ignore build artifacts
+planning_algorithms.egg-info/
